+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v1.1.13 (2020-04-20)
 #### Bug fixes
 - Use `idempotent-babel-polyfill` instead of `babel-polyfill` as a workaround to [this issue](https://github.com/babel/babel-loader/issues/401).
+- Use the `box-sizing` CSS attribute to compute the real with of the counters area in `DefaultSkin`.
 
 #### Improvements
 - Full screen mode now using direct calls to the [Full Screen API](https://developer.mozilla.org/en-US/docs/Web/API/Fullscreen_API) instead of [screenfull](https://www.npmjs.com/package/screenfull).