+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.1.14 (not yet released)
+#### Improvements
+- Upgraded components
+
 ### v1.1.13 (2020-04-20)
 #### Bug fixes
 - Use `idempotent-babel-polyfill` instead of `babel-polyfill` as a workaround to [this issue](https://github.com/babel/babel-loader/issues/401).